Extract `cvd/fetch/download_flags.h`
Bug: b/424021019
Test: bazel test '//...' --test_summary=terse
