+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -63,6 +63,23 @@ clang_tidy_test(
 )
 
 cc_library(
+    name = "download_flags",
+    srcs = ["download_flags.cc"],
+    hdrs = ["download_flags.h"],
+    copts = COPTS,
+    deps = [
+        "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
+        "//cuttlefish/host/commands/cvd/fetch:get_optional",
+    ],
+)
+
+clang_tidy_test(
+    name = "download_flags_clang_tidy",
+    srcs = [":download_flags"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
     name = "downloaders",
     srcs = ["downloaders.cc"],
     hdrs = ["downloaders.h"],
@@ -140,6 +157,7 @@ cc_library(
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/commands/cvd/fetch:de_android_sparse",
+        "//cuttlefish/host/commands/cvd/fetch:download_flags",
         "//cuttlefish/host/commands/cvd/fetch:downloaders",
         "//cuttlefish/host/commands/cvd/fetch:extract_image_contents",
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
