+++ base/cvd/cuttlefish/host/commands/cvd/fetch/download_flags.h
@@ -1,0 +1,29 @@
+//
+// Copyright (C) 2019 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "cuttlefish/host/commands/cvd/fetch/fetch_cvd_parser.h"
+
+namespace cuttlefish {
+
+struct DownloadFlags {
+  static DownloadFlags Create(const VectorFlags& flags, int index);
+
+  bool download_img_zip;
+  bool download_target_files_zip;
+};
+
+}  // namespace cuttlefish
